refactor: split introduction and FAQ into separate components
Remove the combined IntroductionAndStages component and create a new
IntroductionAndFaq component to render the course description and FAQ.
Update templates to use IntroductionAndFaq and Stages separately for
better separation of concerns and improved maintainability.
