+++ app/components/course-overview-page/introduction-and-faq.ts
@@ -1,0 +1,18 @@
+import Component from '@glimmer/component';
+import type CourseModel from 'codecrafters-frontend/models/course';
+
+interface Signature {
+  Element: HTMLDivElement;
+
+  Args: {
+    course: CourseModel;
+  };
+}
+
+export default class CourseOverviewPageIntroductionAndFaq extends Component<Signature> {}
+
+declare module '@glint/environment-ember-loose/registry' {
+  export default interface Registry {
+    'CourseOverviewPage::IntroductionAndFaq': typeof CourseOverviewPageIntroductionAndFaq;
+  }
+}
