+++ app/components/course-overview-page/stages.hbs
@@ -1,16 +1,5 @@
 <div class="bg-white dark:bg-gray-850 rounded-md shadow-xs border border-gray-200 dark:border-white/5 relative" ...attributes>
   <div class="p-4 md:p-6">
-    <div class="leading-7 prose dark:prose-invert mb-4">
-      {{markdown-to-html @course.descriptionMarkdown}}
-    </div>
-
-    {{#if (gt @course.frequentlyAskedQuestions.length 0)}}
-      {{! Extra if convinces typescript that frequentlyAskedQuestions is not null }}
-      {{#if @course.frequentlyAskedQuestions}}
-        <CollapsibleFaqList @faqs={{@course.frequentlyAskedQuestions}} class="mb-6" />
-      {{/if}}
-    {{/if}}
-
     <h3 class="font-semibold text-gray-600 dark:text-gray-400 text-center text-sm relative mb-3">
       <span class="block h-px w-full bg-gray-200 dark:bg-white/5 absolute top-50-percent -mt-px"></span>
       <span class="relative px-6 bg-white dark:bg-gray-850">Stages</span>
