+++ app/components/course-overview-page/stages.ts
@@ -1,0 +1,20 @@
+import Component from '@glimmer/component';
+import type CourseModel from 'codecrafters-frontend/models/course';
+import type CourseStageModel from 'codecrafters-frontend/models/course-stage';
+
+interface Signature {
+  Element: HTMLDivElement;
+
+  Args: {
+    course: CourseModel;
+    completedStages?: CourseStageModel[];
+  };
+}
+
+export default class CourseOverviewPageStages extends Component<Signature> {}
+
+declare module '@glint/environment-ember-loose/registry' {
+  export default interface Registry {
+    'CourseOverviewPage::Stages': typeof CourseOverviewPageStages;
+  }
+}
