+++ app/templates/join-course.hbs
@@ -20,7 +20,8 @@
   <div class="flex items-start flex-row mb-4">
     <div class="grow">
       <CourseOverviewPage::Notices @course={{this.model.course}} class="mb-4" />
-      <CourseOverviewPage::IntroductionAndStages @course={{this.model.course}} class="w-full" />
+      <CourseOverviewPage::IntroductionAndFaq @course={{this.model.course}} class="w-full mb-4" />
+      <CourseOverviewPage::Stages @course={{this.model.course}} class="w-full" />
     </div>
 
     <div class="hidden lg:block sticky top-4 w-80 shrink-0 ml-2">
